# backend/document/openapi.yaml
openapi: 3.0.3
info:
  title: coffee-ranker API
  version: 0.1.0
paths:
...
  /auth/csrf:
    get:
      summary: Issue CSRF token
      tags:
        - auth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CSRFResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /guest-session:
    post:
      summary: Get or create guest session
      tags:
        - guest-session
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestSessionResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: X-Guest-Session-Key
          in: header
          required: false
          schema:
            type: string
  /auth/signup:
    post:
      summary: Signup
      tags:
        - auth
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
  /auth/login:
    post:
      summary: Login
      tags:
        - auth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
  /auth/refresh:
    post:
      summary: Refresh access token
      tags:
        - auth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - csrfHeader: []
          csrfCookie: []
  /auth/logout:
    post:
      summary: Logout current family
      tags:
        - auth
      responses:
        "204":
          description: No Content
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - bearerAuth: []
          csrfHeader: []
          csrfCookie: []
  /auth/logout-all:
    post:
      summary: Logout all devices
      tags:
        - auth
      responses:
        "204":
          description: No Content
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - bearerAuth: []
          csrfHeader: []
          csrfCookie: []
  /auth/me:
    get:
      summary: Me
      tags:
        - auth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - bearerAuth: []
  /beans:
    get:
      summary: List beans
      tags:
        - content
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bean"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
  /beans/{id}:
    get:
      summary: Get bean
      tags:
        - content
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bean"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
  /beans/{id}/articles:
    get:
      summary: List related articles for bean
      tags:
        - content
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BeanArticle"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          required: false
          schema:
            type: integer
  /articles:
    get:
      summary: List articles
      tags:
        - content
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
  /articles/{slug}:
    get:
      summary: Get article by slug
      tags:
        - content
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
  /articles/id/{id}:
    get:
      summary: Get article by id
      tags:
        - content
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      security:
        - bearerAuth: []
  /articles/id/{id}/beans:
    get:
      summary: List related beans for article
      tags:
        - content
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BeanArticle"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      security:
        - bearerAuth: []
  /search/beans:
    get:
      summary: Search beans
      tags:
        - search
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bean"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: origin
          in: query
          required: false
          schema:
            type: string
        - name: roast_level
          in: query
          required: false
          schema:
            type: string
            enum:
              - light
              - medium
              - dark
        - name: acidity
          in: query
          required: false
          schema:
            type: integer
        - name: bitterness
          in: query
          required: false
          schema:
            type: integer
        - name: flavor
          in: query
          required: false
          schema:
            type: integer
        - name: aroma
          in: query
          required: false
          schema:
            type: integer
        - name: body
          in: query
          required: false
          schema:
            type: integer
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum:
              - score
              - newest
              - popular
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      security:
        - {}
        - bearerAuth: []
        - guestSessionCookie: []
  /search/articles:
    get:
      summary: Search articles
      tags:
        - search
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum:
              - brewing
              - roast
              - beans
              - recipe
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum:
              - score
              - newest
              - popular
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      security:
        - {}
        - bearerAuth: []
        - guestSessionCookie: []
  /rankings:
    get:
      summary: List rankings
      tags:
        - ranking
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankingResult"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: content_type
          in: query
          required: false
          schema:
            type: string
            enum:
              - bean
              - article
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
  /rankings/top:
    get:
      summary: Top rankings
      tags:
        - ranking
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentMetric"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
  /recommendations:
    get:
      summary: List recommendations
      tags:
        - recommendation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecommendationItem"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: content_type
          in: query
          required: false
          schema:
            type: string
            enum:
              - bean
              - article
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      security:
        - {}
        - bearerAuth: []
        - guestSessionCookie: []
  /events:
    post:
      summary: Record event
      tags:
        - event
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionEvent"
        "204":
          description: No Content
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventRequest"
      security:
        - csrfHeader: []
          csrfCookie: []
        - bearerAuth: []
          csrfHeader: []
          csrfCookie: []
        - guestSessionCookie: []
          csrfHeader: []
          csrfCookie: []
  /saved:
    get:
      summary: List saved items
      tags:
        - saved
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedItem"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      security:
        - bearerAuth: []
    post:
      summary: Save item
      tags:
        - saved
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedItem"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveRequest"
      security:
        - bearerAuth: []
          csrfHeader: []
          csrfCookie: []
  /saved/{rank_target_id}:
    get:
      summary: Check saved item
      tags:
        - saved
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedExistsResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: rank_target_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      security:
        - bearerAuth: []
    delete:
      summary: Remove saved item
      tags:
        - saved
      responses:
        "204":
          description: No Content
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: rank_target_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      security:
        - bearerAuth: []
          csrfHeader: []
          csrfCookie: []
  /ratings:
    get:
      summary: List ratings
      description: 認証済みUserが付けたGood/Bad評価一覧を返す。
      tags:
        - rating
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rating"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 10000
            default: 0
      security:
        - bearerAuth: []
    post:
      summary: Rate item
      tags:
        - rating
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingRequest"
      security:
        - bearerAuth: []
          csrfHeader: []
          csrfCookie: []
  /ratings/{rank_target_id}:
    get:
      summary: Get rating
      tags:
        - rating
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: rank_target_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      security:
        - bearerAuth: []
    delete:
      summary: Delete rating
      tags:
        - rating
      responses:
        "204":
          description: No Content
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: rank_target_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      security:
        - bearerAuth: []
          csrfHeader: []
          csrfCookie: []
  /modals:
    post:
      summary: Show modal
      tags:
        - modal
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModalDisplayLog"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShowModalRequest"
      security:
        - csrfHeader: []
          csrfCookie: []
        - bearerAuth: []
          csrfHeader: []
          csrfCookie: []
        - guestSessionCookie: []
          csrfHeader: []
          csrfCookie: []
  /modals/click:
    post:
      summary: Click modal
      tags:
        - modal
      responses:
        "204":
          description: No Content
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModalActionRequest"
      security:
        - csrfHeader: []
          csrfCookie: []
        - bearerAuth: []
          csrfHeader: []
          csrfCookie: []
        - guestSessionCookie: []
          csrfHeader: []
          csrfCookie: []
  /modals/close:
    post:
      summary: Close modal
      tags:
        - modal
      responses:
        "204":
          description: No Content
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModalActionRequest"
      security:
        - csrfHeader: []
          csrfCookie: []
        - bearerAuth: []
          csrfHeader: []
          csrfCookie: []
        - guestSessionCookie: []
          csrfHeader: []
          csrfCookie: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
    csrfCookie:
      type: apiKey
      in: cookie
      name: csrf_token
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
    guestSessionCookie:
      type: apiKey
      in: cookie
      name: __Host-guest_session
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    CSRFResponse:
      type: object
      properties:
        csrf_token:
          type: string
    GuestSessionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: boolean
        expires_at:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - user
            - admin
        status:
          type: string
          enum:
            - active
            - suspended
            - deleted
        created_at:
          type: string
        updated_at:
          type: string
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserResponse"
        access_token:
          type: string
    SignupRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
...
    SavedExistsResponse:
      type: object
      properties:
        saved:
          type: boolean
...